Fixed wrong answers bug. Removed rounding the answer.
I fixed a bug that tells the user their answer is wrong when it's actually right. This bug was mainly because of having generated a different set of equations when "Check Answers" is pressed in order to prepare for the later round of equations, and then comparing the user's answers to the newly generated eqautions.

I also removed answer rounding, this improves user experience and main logic: it makes the answers more accurate. Instead, I compare the first 2 decimal places with an error threshold of 0.1